Clean-up, setting up for training implementation #12
