+++ Python_scripts/agent.py
@@ -4,7 +4,7 @@ from cnn import CNN
 
 
 class Agent:
-    def __init__(self, num: int, group, group_specs, vision_decoder: CNN = None) -> None:  # TODO: change vision_decoder
+    def __init__(self, num: int, group, group_specs, vision_decoder: CNN = None) -> None:
         '''
         Initialize the agent with identifiers and its vision decoder.
         '''
@@ -26,13 +26,7 @@ class Agent:
 
         observations_formatted = torch.tensor(observations).permute(2, 0, 1).float()  # Convert to tensor, move channels to first dim
         predictions = self.vision_decoder(observations_formatted)
-        # Make additional computations here in order to cover entire range required (e.g. [-1, 1]) by using the predictions between [0, 1]
         return predictions.detach().numpy()
-
-    def update_decoder(self):
-        # Do a gradient step to update the cnn
-        # Might be better to do that from the trainer directly
-        raise NotImplementedError
     
     def increment_agent_num(self, amount = 1):
         '''
