+++ Python_scripts/experience.py
@@ -1,0 +1,24 @@
+import numpy as np
+from typing import NamedTuple, List
+
+class Experience(NamedTuple):
+    """
+    An experience contains the data of one Agent transition.
+    - Observation
+    - Action
+    - Reward
+    - Done flag
+    - Next Observation
+    """
+
+    obs: np.ndarray
+    action: np.ndarray
+    reward: float
+    done: bool
+    next_obs: np.ndarray
+
+# A Trajectory is an ordered sequence of Experiences
+Trajectory = List[Experience]
+
+# A Buffer is an unordered list of Experiences from multiple Trajectories
+Buffer = List[Experience]
